Refactored the components into separate files
