+++ my-first-react-app/src/hello.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+class Hello extends React.Component {
+  render() {
+    return <h1>Hello {this.props.name}!</h1>;
+  }
+}
+
+
+export default Hello;
